+++ packages/mastodon/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+**Features**
+
+- Add HTTP `Link` header pagination parsing to generated API responses
+
 ## v1.0.0
 
 **Features**